Cleared up 90% of the warnings Clang gives when compiling.

+++ src/monmove.c
@@ -2377,8 +2377,8 @@ struct monst *mtmp;
 
 void
 remove_monster(x, y)
-uchar x;
-uchar y;
+int x;
+int y;
 {
 	struct monst * mtmp = level.monsters[x][y];
